Got code for current week with timedelta from stackoverflow

+++ workloggerapp/models.py
@@ -16,7 +16,7 @@ class Project(models.Model):
 
 class Log(models.Model):
 	project = models.ForeignKey(Project, null=True, on_delete=models.SET_NULL)
-	duration = models.IntegerField()
+	duration = models.DecimalField(max_digits=4, decimal_places=2)
 	remarks = models.CharField(max_length=128)
 	user = models.ForeignKey(User)
 	date = models.DateField( default=datetime.date.today)
